minor update of XeTeX header files
git-svn-id: svn://tug.org/texlive/trunk/Build/source@21519 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/web2c/xetexdir/XeTeXFontInst.h
@@ -57,7 +57,7 @@ authorization from the copyright holders.
 #include "XeTeX_ext.h"
 
 extern "C" {
-	void *xmalloc(unsigned);	// from kpathsea
+	void *xmalloc(size_t);	// from kpathsea
 };
 
 // Abstract superclass that XeTeXOTLayoutEngine uses;
